refactor `tls::rota`
